requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Cache==0.12
Flask-Migrate==1.2.0
Flask-SQLAlchemy==1.0
Flask-Script==0.6.6
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.19
PyYAML==3.10
Werkzeug==0.9.4
anyjson==0.3.3
argparse==1.2.1
gunicorn==18.0
itsdangerous==0.24
lever==0.2.6
decorator
cryptacular==1.4.1
Flask-Login==0.2.11
